requirements.txt
# Pinned versions as of last successful test of this code. Newer versions
# of these packages will most likely work, but haven't been tested.
cfgrib==0.9.10.3
cloudpickle==2.0.0
dask==2023.3.2
distributed==2023.3.2
eccodes==1.5.2
fsspec==2023.4.0
gcsfs==2023.4.0
google-api-core==2.11.0
google-api-python-client==2.57.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-auth==2.17.2
google-cloud-storage==2.5.0
holidays==0.13
joblib==1.2.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter_http_over_ws==0.0.8
kerchunk==0.1.1
keras==2.12.0
matplotlib==3.5.2
mypy==0.950
notebook==6.4.11
nrel-pysam==3.0.2
numba==0.55.2
numcodecs==0.11.0
numpy==1.22.3
openpyxl==3.0.7
pandas==1.4.2
plotly==5.13.0
pvlib==0.9.1
pyarrow==8.0.0
requests==2.27.1
responses==0.22.0
scikit-learn==1.0.2
scipy==1.8.1
shap==0.41.0
statsmodels==0.13.1
tables===3.7.0
tensorflow==2.12.0
ujson==5.4.0
xarray==2023.3.0
xgboost==1.4.2
zarr==2.11.3

# packages that are not required for code execution, but useful for development
ipdb==0.13.9
ipython==8.4.0


